% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.r
\name{begin}
\alias{begin}
\title{Begin generation of data pairs}
\usage{
begin(x, ...)
}
\description{
An SQL statement representing the generation of data pairs, including
the configuration of blocking fields, phonetics etc. is constructed and
send to SQLite.
}
