% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySupv-methods.r
\docType{methods}
\name{classifySupv,RecLinkClassif,RecLinkData-method}
\alias{classifySupv,RecLinkClassif,RecLinkData-method}
\title{Method for old (S3-style) objects, replaces ordinary function}
\usage{
\S4method{classifySupv}{RecLinkClassif,RecLinkData}(model, newdata,
  convert.na = TRUE, ...)
}
\description{
Method for old (S3-style) objects, replaces ordinary function
}
