\name{ff_vector-class}
\Rdversion{1.1}
\docType{class}
\alias{ff_vector-class}

\title{Class \code{"ff_vector"}}
\description{
  S4 representation of S3 class \code{"ff_vector"}, created by \code{\link{setOldClass}}.
  See \code{\link[ff]{ff}} for documentation of the underlying S3 class.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"}}
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
No methods defined with class "ff_vector" in the signature.
}
\keyword{classes}
