\name{getFrequencies-methods}
\docType{methods}
\alias{getFrequencies}
\alias{getFrequencies-methods}
\alias{getFrequencies,RLBigData-method}
\title{Get attribute frequencies}
\description{
  Returns the average frequencies of attribute values for a Record Linkage
  object, which is \code{1 / unique(c)}, for every data column \code{c}.
}
\section{Methods}{
\describe{
  \item{\code{signature(x = "RLBigData")}}{}
}}
\keyword{methods}
\keyword{classif}
