% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.r
\name{getSQLStatement}
\alias{getSQLStatement}
\title{Create SQL statement}
\usage{
getSQLStatement(data1, data2 = data1, con, type, blockFld, excludeFld,
  strcmpFld, strcmpFun, phoneticFld, phoneticFun)
}
\description{
Creates SQL statememt to retreive comparison patterns, respecting
parameters such as blocking definition and exclusion of fields.
}
