\name{getTable-methods}
\docType{methods}
\alias{getTable}
\alias{getTable-methods}
\alias{getTable,RecLinkResult-method}
\alias{getTable,RLResult-method}

\title{Build contingency table}
\description{
  Builds a contingency table for a linkage result with counts for
  each combination of real matching status and predicted result.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "RecLinkResult")}}{
  Method for S3 result sets.
}

\item{\code{signature(object = "RLResult")}}{
  Method for S4 result sets (big data sets).
}
}}
\keyword{methods}
\keyword{classif}
