\name{isFALSE}
\alias{isFALSE}

\title{Check for FALSE}
\description{
  Shortcut for \code{identical(x,FALSE)}
}
\usage{
isFALSE(x)
}

\arguments{
  \item{x}{Any object}
}

\value{
  Returns \code{TRUE} if \code{x} is identical to \code{FALSE}, \code{FALSE}
  otherwise.
}
\author{Andreas Borg}


\keyword{misc}
