\name{texSummary}
\Rdversion{1.1}
\alias{texSummary}
\title{
  LaTeX Summary of linkage results
}
\description{
  LaTeX Summary of linkage results
}
\usage{
texSummary(object)
}

\arguments{
  \item{object}{A RecLinkResultObject}
}

\details{
  For internal usage.
}
\value{
  Used for its side effect.
}

\author{
  Andreas Borg
}


\keyword{internal}